webapp/flask-frruta/fruta-telemetry-server/requirements.txt
Flask==2.2.3
azure-storage-blob==12.10.0
python-dotenv==0.21.0
requests==2.28.1
Jinja2==3.1.2